refactor: add useFilteredTodoList hooks
